refactor(macos): extract playback controls from AppModel into AppModel+PlaybackControls.swift
Continues the AppModel god-object split. Relocates the playback transport +
output-control band — play/pause/stop/skip/volume, relative & absolute seek,
togglePlayPause, gapless pre-load arming, Core Audio output-device routing +
exclusive (hog) mode, ReplayGain normalization, and the gapless / stop-after-
current preferences — into a new `extension AppModel` file as one contiguous
block.

Pure relocation: the 290 moved lines are byte-identical to what was removed.
AppModel.swift drops 5,872 -> 5,581 lines. The band has no instance stored
properties — only methods, static UserDefaults keys (which can live in an
extension), and computed props — so nothing stayed behind.

Three access promotions, all band-internal privates called from the
Now-Playing / status-polling / play methods that remain on AppModel for now:
playCurrent, armNextTrackPreload (private -> internal) and resetStopAfterCurrent
(private static -> internal static). nextTrackForPreload stays private (only
used within the band); consumeStopAfterCurrent was already internal.

Verified: clean rebuild (rm -rf macos/.build && swift build) + swift test
(889 tests) green.
